+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+<a name="v2.28.0"></a>
+## [v2.28.0] - 2024-04-02
+### Bug Fixes
+- **testhelpers:** add insights API mock URL to testconfig ([#1077](https://github.com/newrelic/newrelic-client-go/issues/1077))
+
+### Features
+- **entities:** add entity search query helpers
+
+<a name="v2.27.0"></a>
+## [v2.27.0] - 2024-03-20
+### Bug Fixes
+- Logger string interpolation ([#1104](https://github.com/newrelic/newrelic-client-go/issues/1104))
+- **nrdb:** update to nrdb query to include more options ([#1110](https://github.com/newrelic/newrelic-client-go/issues/1110))
+
+### Features
+- synthetic monitors `MUTED` status EOL in newrelic-client-go ([#1102](https://github.com/newrelic/newrelic-client-go/issues/1102))
+- **dashboards:** Added ignore time picker as options field to the dashboard variables ([#1106](https://github.com/newrelic/newrelic-client-go/issues/1106))
+- **synthetics:** addition of support for next-gen runtime to step, cert check, broken links monitors ([#1114](https://github.com/newrelic/newrelic-client-go/issues/1114))
+
 <a name="v2.26.1"></a>
 ## [v2.26.1] - 2024-03-01
 ### Bug Fixes
@@ -1712,7 +1731,9 @@
 - extract paging implementation
 - rename packages for clarity, promote Config to the public package
 
-[Unreleased]: https://github.com/newrelic/newrelic-client-go/compare/v2.26.1...HEAD
+[Unreleased]: https://github.com/newrelic/newrelic-client-go/compare/v2.28.0...HEAD
+[v2.28.0]: https://github.com/newrelic/newrelic-client-go/compare/v2.27.0...v2.28.0
+[v2.27.0]: https://github.com/newrelic/newrelic-client-go/compare/v2.26.1...v2.27.0
 [v2.26.1]: https://github.com/newrelic/newrelic-client-go/compare/v2.26.0...v2.26.1
 [v2.26.0]: https://github.com/newrelic/newrelic-client-go/compare/v2.25.0...v2.26.0
 [v2.25.0]: https://github.com/newrelic/newrelic-client-go/compare/v2.24.0...v2.25.0